Consolidate `Regions` and UI for `Field Sizes`
