+++ webview/src/routes/regions/+page.svelte
@@ -1,34 +1,15 @@
 <script lang="ts">
-	import { type Region } from '$lib';
 	import { slide } from 'svelte/transition';
 	import RegionList from './RegionList.svelte';
-	import { getModalStore, getToastStore } from '@skeletonlabs/skeleton';
-
-	const toastStore = getToastStore();
-	const modalStore = getModalStore();
+	import CreateRegionButton from './CreateRegionButton.svelte';
 
 	let regionList: RegionList;
-
-	function createRegion() {
-		modalStore.trigger({
-			type: 'component',
-			component: 'regionCreate',
-			meta: {
-				onCreate(region: Region) {
-					toastStore.trigger({
-						message: `Created new region: "${region.title}"`,
-						background: 'variant-filled-success'
-					});
-					regionList.addRegionToFrontend(region);
-				}
-			}
-		});
-	}
 </script>
 
 <main in:slide={{ axis: 'x' }} out:slide={{ axis: 'x' }} class="p-4">
 	<h2 class="h2">Regions</h2>
 
 	<RegionList bind:this={regionList} />
-	<button class="variant-filled btn mx-auto block" on:click={createRegion}> Create Region </button>
+
+	<CreateRegionButton {regionList} />
 </main>
